# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Qorix Markets API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/deposit:
    post:
      operationId: deposit
      tags: [wallet]
      summary: Deposit USDT to main balance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositBody"
      responses:
        "200":
          description: Deposit processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /wallet/withdraw:
    post:
      operationId: withdraw
      tags: [wallet]
      summary: Request withdrawal from profit balance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawBody"
      responses:
        "200":
          description: Withdrawal requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Insufficient balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /wallet/transfer:
    post:
      operationId: transferToTrading
      tags: [wallet]
      summary: Transfer from main to trading balance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferBody"
      responses:
        "200":
          description: Transfer done
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"

  # TRANSACTIONS
...
  /investment/start:
    post:
      operationId: startInvestment
      tags: [investment]
      summary: Start auto trading investment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartInvestmentBody"
      responses:
        "200":
          description: Investment started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /investment/compounding:
    patch:
      operationId: toggleCompounding
      tags: [investment]
      summary: Toggle auto compounding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompoundingBody"
      responses:
        "200":
          description: Compounding toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"

  # TRADES (simulated)
  /trades:
    get:
      operationId: getTrades
      tags: [investment]
      summary: Get recent simulated trades
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"

  # REFERRAL
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # AUTH
...
    Wallet:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        mainBalance:
          type: number
        tradingBalance:
          type: number
        profitBalance:
          type: number
        updatedAt:
          type: string
      required:
        - id
        - userId
        - mainBalance
        - tradingBalance
        - profitBalance
        - updatedAt
    DepositBody:
      type: object
      properties:
        amount:
          type: number
        txHash:
          type: ["string", "null"]
      required:
        - amount
    WithdrawBody:
      type: object
      properties:
        amount:
          type: number
        walletAddress:
          type: string
      required:
        - amount
        - walletAddress
    TransferBody:
      type: object
      properties:
        amount:
          type: number
      required:
        - amount

    # TRANSACTIONS
    Transaction:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        amount:
          type: number
        status:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - userId
        - type
        - amount
        - status
        - createdAt
...
    Investment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        riskLevel:
          type: string
        isActive:
          type: boolean
        autoCompound:
          type: boolean
        totalProfit:
          type: number
        dailyProfit:
          type: number
        drawdown:
          type: number
        startedAt:
          type: ["string", "null"]
        stoppedAt:
          type: ["string", "null"]
      required:
        - id
        - userId
        - amount
        - riskLevel
        - isActive
        - autoCompound
        - totalProfit
        - dailyProfit
        - drawdown
    StartInvestmentBody:
      type: object
      properties:
        amount:
          type: number
        riskLevel:
          type: string
      required:
        - amount
        - riskLevel
    CompoundingBody:
      type: object
      properties:
        autoCompound:
          type: boolean
      required:
        - autoCompound
    Trade:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        direction:
          type: string
        entryPrice:
          type: number
        exitPrice:
          type: number
        profit:
          type: number
        profitPercent:
          type: number
        executedAt:
          type: string
      required:
        - id
        - symbol
        - direction
        - entryPrice
        - exitPrice
        - profit
        - profitPercent
        - executedAt

    # REFERRAL
...