# lib/api-zod/src/generated/types/investment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Qorix Markets API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Investment {
  id: number;
  userId: number;
  amount: number;
  riskLevel: string;
  isActive: boolean;
  autoCompound: boolean;
  totalProfit: number;
  dailyProfit: number;
  drawdown: number;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  stoppedAt?: string | null;
}