# lib/api-zod/src/generated/types/withdrawBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Qorix Markets API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WithdrawBody {
  amount: number;
  walletAddress: string;
}